change: Require password length to be at least 12 characters instead of only 6
TODO: Add validation to Firebase that only allows accounts to be created
if the password provided is 12 characters or greater. Not sure if this
is possible, but if so, then it would be more secure, although not too
useful in practicality since password length requirements primarily
protect legitimate users and not those who would
circumvent the client side
form's requirements.